Title: 20대의 평균나이, Time: 210ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240903/20대의 평균나이/average-age-of-20.java
@@ -16,6 +16,6 @@ public class Main {
             }
             
         }
-        System.out.printf("%.2f",s/c);
+        System.out.printf("%.2f",(double)s/c);
     }
 }
